# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nursing App API Documentation",
        "description": "API documentation for the Nursing App - A platform connecting users with nursing services",
        "contact": {
            "name": "API Support",
            "email": "support@nursingapp.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/about": {
            "get": {
                "tags": [
                    "About"
                ],
                "summary": "Get about information",
                "description": "Retrieve the about us information. Available to both users and admins.",
                "operationId": "2c4f4365b3953343e1ee44c84ef4538e",
                "responses": {
                    "200": {
                        "description": "About information retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "about": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "online_shop_url": {
                                                    "type": "string",
                                                    "example": "https://shop.example.com"
                                                },
                                                "facebook_url": {
                                                    "type": "string",
                                                    "example": "https://facebook.com/company"
                                                },
                                                "instagram_url": {
                                                    "type": "string",
                                                    "example": "https://instagram.com/company"
                                                },
                                                "tiktok_url": {
                                                    "type": "string",
                                                    "example": "https://tiktok.com/@company"
                                                },
                                                "whatsapp_numbers": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "+1234567890",
                                                        "+0987654321"
                                                    ]
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "We provide professional nursing care services"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/about": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update about information (Admin only)",
                "description": "Update the about us information. Only accessible by admins.",
                "operationId": "c09c2e5f10642d6e5eeb5a44afee33df",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "online_shop_url": {
                                        "description": "Online shop URL",
                                        "type": "string",
                                        "example": "https://shop.example.com"
                                    },
                                    "facebook_url": {
                                        "description": "Facebook page URL",
                                        "type": "string",
                                        "example": "https://facebook.com/company"
                                    },
                                    "instagram_url": {
                                        "description": "Instagram profile URL",
                                        "type": "string",
                                        "example": "https://instagram.com/company"
                                    },
                                    "tiktok_url": {
                                        "description": "TikTok profile URL",
                                        "type": "string",
                                        "example": "https://tiktok.com/@company"
                                    },
                                    "whatsapp_numbers": {
                                        "description": "Array of WhatsApp numbers",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "+1234567890",
                                            "+0987654321"
                                        ]
                                    },
                                    "description": {
                                        "description": "About description",
                                        "type": "string",
                                        "example": "We provide professional nursing care services"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "About information updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "About Us updated successfully."
                                        },
                                        "about": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "online_shop_url": {
                                                    "type": "string",
                                                    "example": "https://shop.example.com"
                                                },
                                                "facebook_url": {
                                                    "type": "string",
                                                    "example": "https://facebook.com/company"
                                                },
                                                "instagram_url": {
                                                    "type": "string",
                                                    "example": "https://instagram.com/company"
                                                },
                                                "tiktok_url": {
                                                    "type": "string",
                                                    "example": "https://tiktok.com/@company"
                                                },
                                                "whatsapp_numbers": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "+1234567890",
                                                        "+0987654321"
                                                    ]
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "We provide professional nursing care services"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin role required"
                    },
                    "404": {
                        "description": "About information not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}